==> code/analysis_final/1_clinical_data_summary.Rmd <==
---
title: "Clinical data summary stats"
author: "Leah Lariscy"
output: html_document
---

# Load packages

```{r}
library(tidyverse)
library(tsibble)
library(here)
```

# Load data

```{r}
weekly <- readRDS(
  here("data/processed_data/clinical/covid_cases_weekly.rds")) %>% 
  mutate(date = as.Date(week, format = "%Y-%m-%d"))
```

# Summarize

## Total reported cases

```{r}
sum(weekly$cases.reported)
```

## Highest case weeks

```{r}
weekly %>% arrange(desc(cases.reported))
```

## Lowest case weeks

```{r}
weekly %>% arrange(cases.reported)
```

## Highest test weeks

```{r}
weekly %>% arrange(desc(pcr_tests))
```

## Lowest test weeks

```{r}
weekly %>% arrange(pcr_tests)
```

## Stats

```{r}
weekly_long <- weekly %>% select(cases.reported, pcr_tests) %>% 
  rename("Reported Cases" = cases.reported,
         "Tests Administered" = pcr_tests) %>% 
  pivot_longer(cols = c("Reported Cases","Tests Administered"),
              names_to = "Clinical Metric")

summary <- weekly_long %>% group_by(`Clinical Metric`) %>% 
  summarise(Mean = mean(value),
            Median = median(value),
            SD = sd(value),
            Min = min(value),
            Max = max(value))

write_csv(summary, here("data/processed_data/clinical/summary_stats.csv"))
```

# Visualize

```{r}
cases_reported_fig <- weekly %>% ggplot(aes(week,cases.reported)) +
  geom_point(color = "#FB6A4A", size = 2) +
  geom_line(alpha = .6, color = "#FB6A4A", linewidth = 2) +
  geom_vline(xintercept = 18974,linetype = 2, size = .8) +
  #geom_area(alpha = .7, fill = "#FB6A4A") +
  ylab("Cases reported") +
  xlab("") +
  ggthemes::theme_clean()

cases_reported_fig
```

```{r}
test_admin_fig <- weekly %>% ggplot(aes(week,pcr_tests)) +
  geom_point(color = "#A50F15", size = 2) +
  geom_line(alpha = .6, color = "#A50F15", linewidth = 2) +
  geom_vline(xintercept = 18974,linetype = 2, size = .8) +
  #geom_area(alpha = .7, fill = "#A50F15") +
  ylab("Tests administered") +
  xlab("") +
  ggthemes::theme_clean()

test_admin_fig
```

```{r}
ggarrange(cases_reported_fig,test_admin_fig,
          nrow = 2,
          ncol = 1,
          labels = c("a","b"))

ggsave(here("figures/timeseries/clinical_combined.tiff"),
       width = 9,
       height = 5)
```
