==> code/analysis_final/3_corr_matrix.Rmd <==
---
title: "Correlation matrix"
author: "Leah Lariscy"
output: html_document
---

# Load packages

```{r}
library(tidyverse)
library(corrplot)
library(tsibble)
library(reshape2)
library(here)
library(ggpubr)
```

# Load data

```{r}
data <- readRDS(here("data/processed_data/wbe_covid_n6_week.rds"))
```

# Create matrix

## Full data

```{r}
correlations <- data %>%
  #head(n=75) %>% 
  drop_na() %>% 
  select('Plant A N1 VL'=A_N1, 'Plant A N2 VL'=A_N2,
         'Plant B N1 VL'=B_N1, 'Plant B N2 VL'=B_N2,
         'Plant C N1 VL'=C_N1, 'Plant C N2 VL'=C_N2,
         'Plant A N1 DF'=A_N1_POS, 'Plant A N2 DF'=A_N2_POS,
         'Plant B N1 DF'=B_N1_POS, 'Plant B N2 DF'=B_N2_POS,
         'Plant C N1 DF'=C_N1_POS, 'Plant C N2 DF'=C_N2_POS,
         'Cases Reported'=cases.reported) %>% 
  drop_na() %>% 
  cor(method = "spearman")
```

## Training data

```{r}
correlations_train <- data %>%
  head(n=75) %>% 
  drop_na() %>% 
  select('Plant A N1 VL'=A_N1, 'Plant A N2 VL'=A_N2,
         'Plant B N1 VL'=B_N1, 'Plant B N2 VL'=B_N2,
         'Plant C N1 VL'=C_N1, 'Plant C N2 VL'=C_N2,
         'Plant A N1 DF'=A_N1_POS, 'Plant A N2 DF'=A_N2_POS,
         'Plant B N1 DF'=B_N1_POS, 'Plant B N2 DF'=B_N2_POS,
         'Plant C N1 DF'=C_N1_POS, 'Plant C N2 DF'=C_N2_POS,
         'Cases Reported'=cases.reported) %>% 
  drop_na() %>% 
  cor(method = "spearman")
```

## Testing data

```{r}
correlations_test <- data %>%
  tail(n=52) %>% 
  drop_na() %>% 
  select('Plant A N1 VL'=A_N1, 'Plant A N2 VL'=A_N2,
         'Plant B N1 VL'=B_N1, 'Plant B N2 VL'=B_N2,
         'Plant C N1 VL'=C_N1, 'Plant C N2 VL'=C_N2,
         'Plant A N1 DF'=A_N1_POS, 'Plant A N2 DF'=A_N2_POS,
         'Plant B N1 DF'=B_N1_POS, 'Plant B N2 DF'=B_N2_POS,
         'Plant C N1 DF'=C_N1_POS, 'Plant C N2 DF'=C_N2_POS,
         'Cases Reported'=cases.reported) %>% 
  drop_na() %>% 
  cor(method = "spearman")
```

## Year 2+

```{r}
correlations_y2 <- data %>%
  tail(n=78) %>% 
  drop_na() %>% 
  select('Plant A N1 VL'=A_N1, 'Plant A N2 VL'=A_N2,
         'Plant B N1 VL'=B_N1, 'Plant B N2 VL'=B_N2,
         'Plant C N1 VL'=C_N1, 'Plant C N2 VL'=C_N2,
         'Plant A N1 DF'=A_N1_POS, 'Plant A N2 DF'=A_N2_POS,
         'Plant B N1 DF'=B_N1_POS, 'Plant B N2 DF'=B_N2_POS,
         'Plant C N1 DF'=C_N1_POS, 'Plant C N2 DF'=C_N2_POS,
         'Cases Reported'=cases.reported) %>% 
  drop_na() %>% 
  cor(method = "spearman")
```

## Year 1

```{r}
correlations_y1 <- data %>%
  head(53) %>% 
  drop_na() %>% 
  select('Plant A N1 VL'=A_N1, 'Plant A N2 VL'=A_N2,
         'Plant B N1 VL'=B_N1, 'Plant B N2 VL'=B_N2,
         'Plant C N1 VL'=C_N1, 'Plant C N2 VL'=C_N2,
         'Plant A N1 DF'=A_N1_POS, 'Plant A N2 DF'=A_N2_POS,
         'Plant B N1 DF'=B_N1_POS, 'Plant B N2 DF'=B_N2_POS,
         'Plant C N1 DF'=C_N1_POS, 'Plant C N2 DF'=C_N2_POS,
         'Cases Reported'=cases.reported) %>% 
  drop_na() %>% 
  cor(method = "spearman")
```

# Plot matrix

## Full data

```{r}
# Reshape the correlation matrix using melt
cor_matrix_melted <- melt(correlations)

#filter out duplicate correlations
cor_matrix_melted <- cor_matrix_melted %>%
  filter(as.numeric(Var1) > as.numeric(Var2))

# Create the heatmap with numerical values
ggplot(cor_matrix_melted, aes(x = Var1, y = Var2, fill = value)) +
  geom_tile(color = "white") +
  scale_fill_gradient2(low = "blue", high = "red", mid = "white",
                       midpoint = .2, limit = c(-.5, 1), space = "Lab",
                       name="Spearman\nCorrelation") +
  geom_text(aes(label = round(value, 2)), color = "black", size = 2.5) +
  ggthemes::theme_clean() + 
  theme(axis.text.x = element_text(angle = 45, vjust = 1, 
                                   hjust = 1),
        axis.title = element_blank()) +
  coord_fixed()

ggsave(here("figures/corr_matrix/spearman_full.tiff"))
```

## Training data

```{r}
# Reshape the correlation matrix using melt
cor_matrix_melted_train <- melt(correlations_train)

#filter out duplicate correlations
cor_matrix_melted_train <- cor_matrix_melted_train %>%
  filter(as.numeric(Var1) > as.numeric(Var2))

# Create the heatmap with numerical values
ggplot(cor_matrix_melted_train, aes(x = Var1, y = Var2, fill = value)) +
  geom_tile(color = "white") +
  scale_fill_gradient2(low = "blue", high = "red", mid = "white",
                       midpoint = .2, limit = c(-.5, 1), space = "Lab",
                       name="Spearman\nCorrelation") +
  geom_text(aes(label = round(value, 2)), color = "black", size = 2.5) +
  ggthemes::theme_clean() + 
  theme(axis.text.x = element_text(angle = 45, vjust = 1, 
                                   hjust = 1),
        axis.title = element_blank()) +
  coord_fixed()

ggsave(here("figures/corr_matrix/spearman_training.tiff"))
```

## Year 2+

```{r}
# Reshape the correlation matrix using melt
cor_matrix_melted_y2 <- melt(correlations_y2)

#filter out duplicate correlations
cor_matrix_melted_y2 <- cor_matrix_melted_y2 %>%
  filter(as.numeric(Var1) > as.numeric(Var2))

# Create the heatmap with numerical values
y2_matrix <- ggplot(cor_matrix_melted_y2, aes(x = Var1, y = Var2, fill = value)) +
  geom_tile(color = "white") +
  scale_fill_gradient2(low = "blue", high = "red", mid = "white",
                       midpoint = .2, limit = c(-0.5, 1), space = "Lab",
                       name="Rho",
                      ) +
  geom_text(aes(label = round(value, 2)), color = "black", size = 2.5) +
  ggthemes::theme_clean() + 
  theme(axis.text.x = element_text(angle = 45, vjust = 1, 
                                   hjust = 1),
        axis.title = element_blank()) +
  coord_fixed()

ggsave(here("figures/corr_matrix/spearman_year2.tiff"))
```

## Year 1

```{r}
# Reshape the correlation matrix using melt
cor_matrix_melted_y1 <- melt(correlations_y1)

#filter out duplicate correlations
cor_matrix_melted_y1 <- cor_matrix_melted_y1 %>%
  filter(as.numeric(Var1) > as.numeric(Var2))

# Create the heatmap with numerical values
y1_matrix <- ggplot(cor_matrix_melted_y1, aes(x = Var1, y = Var2, fill = value)) +
  geom_tile(color = "white") +
  scale_fill_gradient2(low = "blue", high = "red", mid = "white",
                       midpoint = .2, limit = c(-.5, 1), space = "Lab",
                       name="Rho") +
  geom_text(aes(label = round(value, 2)), color = "black", size = 2.5) +
  ggthemes::theme_clean() + 
  theme(axis.text.x = element_text(angle = 45, vjust = 1, 
                                   hjust = 1),
        axis.title = element_blank()) +
  coord_fixed()

ggsave(here("figures/corr_matrix/spearman_year1.tiff"))
```

```{r}
ggarrange(y1_matrix, y2_matrix,
          ncol = 2,
          nrow = 1,
          labels = c("a","b"),
          vjust = 37)

ggsave(here("figures/corr_matrix/y1_y2_combo.tiff"),
       width = 9,
       height = 7)
```
