==> code/analysis_final/11_random_forest_compare.Rmd <==
---
title: "Random Forest Comparisons"
author: "Leah Lariscy"
output: html_document
---

# Load packages

```{r}
library(tidyverse)
library(here)
library(RColorBrewer)
library(vip)
library(tsibble)
library(ggpubr)
library(yardstick)
```

# Load data

```{r}
#detection frequency metrics
df_metrics <- readRDS(here("data/processed_data/rand_forest/cv_metrics_df_sum.rds"))

#viral load metrics
vl_metrics <- readRDS(here("data/processed_data/rand_forest/cv_metrics_vl_sum.rds"))

#all predictors metrics
all_metrics <- readRDS(here("data/processed_data/rand_forest/cv_metrics_all_sum.rds"))

#bind metrics dataframes
metrics_compare <- rbind(df_metrics,
                         vl_metrics,
                         all_metrics) %>% 
  mutate(config = paste(feature,n,sep = "_"))

saveRDS(metrics_compare,
        here("data/processed_data/rand_forest/metrics_compare.rds"))

#detection frequency n=3 predictions and rsq/rmse
df_n3_train_predictions <- readRDS(here(
  "data/processed_data/rand_forest/df_n3_train_predictions.rds"))
df_n3_train_rsq <- readRDS(here(
  "data/processed_data/rand_forest/df_n3_train_rsq.rds"))
df_n3_train_rmse <- readRDS(here(
  "data/processed_data/rand_forest/df_n3_train_rmse.rds"))

df_n2_test_predictions <- readRDS(here(
  "data/processed_data/rand_forest/df_n2_test_predictions.rds"))
df_n2_test_rsq <- readRDS(here(
  "data/processed_data/rand_forest/df_n2_test_rsq.rds"))
df_n2_test_rmse <- readRDS(here(
  "data/processed_data/rand_forest/df_n2_test_rmse.rds"))


#all predictors n=6 predictions and rsq/rmse
all_n6_train_predictions <- readRDS(here(
  "data/processed_data/rand_forest/all_n6_train_predictions.rds"))
all_n6_train_rsq <- readRDS(here(
  "data/processed_data/rand_forest/all_n6_train_rsq.rds"))
all_n6_train_rmse <- readRDS(here(
  "data/processed_data/rand_forest/all_n6_train_rmse.rds"))

all_n2_test_predictions <- readRDS(here(
  "data/processed_data/rand_forest/all_n2_test_predictions.rds"))
all_n2_test_rsq <- readRDS(here(
  "data/processed_data/rand_forest/all_n2_test_rsq.rds"))
all_n2_test_rmse <- readRDS(here(
  "data/processed_data/rand_forest/all_n2_test_rmse.rds"))

#all predictors n=6 fit (for variable importance viz)
all_n6_train_fit <- readRDS(
  here("data/processed_data/rand_forest/ALL_n6_fit_train.rds"))
df_n6_train_fit <- readRDS(
  here("data/processed_data/rand_forest/DF_n6_fit_train.rds"))
vl_n6_train_fit <- readRDS(
  here("data/processed_data/rand_forest/VL_n6_fit_train.rds"))
```

# Calculate RMSE/RSQ

```{r}
#RSQ
#calculate for beginning of 2022 (January - April 2022)

#DF 
start_2022_rsq_df <- df_n2_test_predictions %>% 
  head(17) %>% 
  rsq(truth = actual, estimate = prediction)

start_2022_rsq_df

end_2022_rsq_df <- df_n2_test_predictions %>% 
  tail(33) %>% 
  rsq(truth = actual, estimate = prediction)

end_2022_rsq_df

#ALL
start_2022_rsq_all <- all_n2_test_predictions %>% 
  head(17) %>% 
  rsq(truth = actual, estimate = prediction)

start_2022_rsq_all

end_2022_rsq_all <- all_n2_test_predictions %>% 
  tail(33) %>% 
  rsq(truth = actual, estimate = prediction)

end_2022_rsq_all
```

```{r}
#RMSE
#calculate for beginning of 2022 (January - April 2022)

#DF 
start_2022_rmse_df <- df_n2_test_predictions %>% 
  head(17) %>% 
  rmse(truth = actual, estimate = prediction)

start_2022_rmse_df

end_2022_rmse_df <- df_n2_test_predictions %>% 
  tail(33) %>% 
  rmse(truth = actual, estimate = prediction)

end_2022_rmse_df

#ALL
start_2022_rmse_all <- all_n2_test_predictions %>% 
  head(17) %>% 
  rmse(truth = actual, estimate = prediction)

start_2022_rmse_all

end_2022_rmse_all <- all_n2_test_predictions %>% 
  tail(33) %>% 
  rmse(truth = actual, estimate = prediction)

end_2022_rmse_all
```

# Visualize

## Make color palette

```{r}
# Extract the last 6 colors from the "Reds", "Greens", and "Blues" palettes
reds <- brewer.pal(9, "Reds")[4:9]
greens <- brewer.pal(9, "Greens")[4:9]
blues <- brewer.pal(9, "Blues")[4:9]

# Combine the colors into a single list
color_palette <- c(reds, greens, blues)

# Display the colors to confirm their distinctiveness
print(color_palette)

# Ensure the list has exactly 18 colors
color_palette <- color_palette[1:18]

# Make list for naming models in ggplot
labels <- c("All 1","All 2","All 3","All 4","All 5","All 6",
            "DF 1","DF 2","DF 3","DF 4","DF 5","DF 6",
            "VL 1","VL 2","VL 3","VL 4","VL 5","VL 6")
```

## RMSE

```{r}
rmse_compare <- metrics_compare %>% 
  ggplot(aes(config, avg_rmse, color=config)) +
  geom_point(size = 4) +
    geom_errorbar(aes(ymin = lower_rmse,
                ymax = upper_rmse)) +
  ggthemes::theme_clean() +
  theme(axis.title.x = element_blank(),
        #axis.ticks.x = element_blank(),
        axis.text.x = element_text(angle = 20),
        axis.title.y = element_text(size = 12,face = "bold"),
        legend.text = element_text(size = 10),
        legend.title = element_text(size = 12),
        legend.key.height = unit(0.6, "cm"),
        legend.position = "none") + #no legend after all..
  scale_color_manual(name="Configuration",
                    #labels=labels,
                    values = color_palette) +
  xlab("Feature selection configuration") +
  ylab("RMSE") +
  scale_x_discrete(labels=labels) +
  geom_hline(yintercept = 200, alpha = 0) +
  geom_hline(yintercept = 90, alpha = 0) +
      geom_hline(yintercept = 174,
             linetype = "longdash",
             linewidth = .8,
             color = "red") +
  annotate("text", label="Null model average", 
           x=16, y=180, size=4, color = "red")

rmse_compare

ggsave(here("figures/random_forest/metrics/rmse_compare.png"), width = 9, height = 5)

saveRDS(rmse_compare, 
        here("figures/random_forest/metrics/rmse_compare.rds"))
```

```{r}
rmse <- metrics_compare %>% 
  select(model,feature,n,avg_rmse,se_rmse) %>% 
  arrange(avg_rmse)

rmse

saveRDS(rmse,
        here("data/processed_data/rand_forest/rmse_compare.rds"))
```

## RSQ

```{r}
rsq_compare <- metrics_compare %>% 
  ggplot(aes(config, avg_rsq, color=config)) +
  geom_point(size = 4) +
    geom_errorbar(aes(ymin = lower_rsq,
                ymax = upper_rsq)) +
  ggthemes::theme_clean() +
  theme(axis.title.x = element_blank(),
        #axis.ticks.x = element_blank(),
        axis.text.x = element_text(angle = 20),
        axis.title = element_text(size = 12),
        legend.text = element_text(size = 10),
        legend.title = element_text(size = 12),
        legend.key.height = unit(0.6, "cm"),
        legend.position = "none") + #no legend after all..
  scale_color_manual(name="Configuration",
                    #labels=labels,
                    values = color_palette) +
  #xlab("Feature selection configuration") +
  ylab("R-Squared") +
  scale_x_discrete(labels=labels) +
  geom_hline(yintercept = 0.8, alpha = 0) +
  geom_hline(yintercept = 0.07,
             linetype = "longdash",
             linewidth = .8,
             color = "red") +
  annotate("text", label="Null model average", 
           x=16, y=.1, size=4, color = "red")

rsq_compare

ggsave(here("figures/random_forest/metrics/rsq_compare.png"), width = 9, height = 5)

saveRDS(rsq_compare,
        here("figures/random_forest/metrics/rsq_compare.rds"))
```

```{r}
rsq <- metrics_compare %>% 
  select(model,feature,n,avg_rsq,se_rsq) %>% 
  arrange(desc(avg_rsq))

rsq

saveRDS(rsq,
        here("data/processed_data/rand_forest/rsq_compare.rds"))
```

## Predictions

### All predictors, n = 6

#### Training data (2020-2021)

```{r}
all_n6_train_rsq_round <- all_n6_train_rsq$.estimate %>% round(digits = 3)

all_n6_train_fig <- all_n6_train_predictions %>% 
  ggplot(aes(x=week)) +
  geom_line(aes(y=prediction), color = "#FB6A4A") +
  geom_point(aes(y=prediction), color = "#FB6A4A", shape = 17) +
  geom_ribbon(aes(ymin=pred_lower, ymax=pred_upper), fill = "#FB6A4A", alpha = 0.3) +
  geom_line(aes(y=actual), color = "#003e99") +
  geom_point(aes(y=actual), color = "#003e99", shape = 16) +
  ggthemes::theme_clean() +
  theme(axis.title = element_text(face = "bold",size = 12),
        axis.title.x = element_blank()) 

all_n6_train_fig

ggsave(here("figures/random_forest/time_series/all_n6_train.png"))

saveRDS(all_n6_train_fig,
        here("figures/random_forest/time_series/all_n6_train.rds"))
```

#### New data (2022)

```{r}
all_n2_test1_rsq_round <- start_2022_rsq_all$.estimate %>% 
  round(digits = 3)
all_n2_test2_rsq_round <- end_2022_rsq_all$.estimate %>% 
  round(digits = 3)
all_n2_test1_rmse_round <- start_2022_rmse_all$.estimate %>% 
  round(digits = 3)
all_n2_test2_rmse_round <- end_2022_rmse_all$.estimate %>% 
  round(digits = 3)

all_n2_test_fig <- all_n2_test_predictions %>% 
  ggplot(aes(x=week)) +
  geom_line(aes(y=prediction), color = "#FB6A4A",
            linewidth = 2, alpha = .5) +
  geom_point(aes(y=prediction), color = "#FB6A4A", 
             shape = 17, size = 2) +
  geom_ribbon(aes(ymin=pred_lower, ymax=pred_upper), fill = "#FB6A4A", alpha = 0.3) +
  geom_line(aes(y=actual), color = "#003e99",
            linewidth = 2, alpha = .5) +
  geom_point(aes(y=actual), color = "#003e99",
             size = 2) +
  annotate("text", label=all_n2_test1_rsq_round, x=19095, y=1500, size=4) +
  annotate("text", label=all_n2_test2_rsq_round, x=19330, y=1500, size=4) +
  annotate("text", label=all_n2_test1_rmse_round, x=19095, y=1600, size=4) +
  annotate("text", label=all_n2_test2_rmse_round, x=19330, y=1600, size=4) +
  annotate("text", label="Predicted", x=19045, y=1425, color="#FB6A4A", size=4) +
  annotate("text", label="Observed", x=19045, y=1500, color="#003e99", size=4) +
  ylab("Cases") +
  #xlab("Week") +
  geom_vline(xintercept = 19115, linetype = 2) +
  #eom_hline(yintercept = 6800, alpha = 0) +
  ggthemes::theme_clean() +
  theme(axis.title = element_text(size = 12),
        axis.title.x = element_blank()) 

all_n2_test_fig

ggsave(here("figures/random_forest/time_series/all_n2_test.png"))

saveRDS(all_n2_test_fig,
        here("figures/random_forest/time_series/all_n2_test.rds"))
```

### Detection frequency, n = 3

#### Training data (2020-2021)

```{r}
df_n3_train_rsq_round <- df_n3_train_rsq$.estimate %>% round(digits = 3)

df_n3_train_fig <- df_n3_train_predictions %>% 
  ggplot(aes(x=week)) +
  geom_line(aes(y=prediction), color = "#FB6A4A") +
  geom_point(aes(y=prediction), color = "#FB6A4A", shape = 17) +
  geom_ribbon(aes(ymin=pred_lower, ymax=pred_upper), fill = "#FB6A4A", alpha = 0.3) +
  geom_line(aes(y=actual), color = "#003e99") +
  geom_point(aes(y=actual), color = "#003e99", shape = 16) +
  ggthemes::theme_clean() +
  theme(axis.title.x = element_blank(),
        axis.title.y = element_blank()) 

df_n3_train_fig

ggsave(here("figures/random_forest/time_series/df_n3_train.png"))

saveRDS(df_n3_train_fig,
        here("figures/random_forest/time_series/df_n3_train.rds"))
```

#### New data (2022)

```{r}
df_n2_test1_rsq_round <- start_2022_rsq_df$.estimate %>% 
  round(digits = 3)
df_n2_test2_rsq_round <- end_2022_rsq_df$.estimate %>% 
  round(digits = 3)
df_n2_test1_rmse_round <- start_2022_rmse_df$.estimate %>% 
  round(digits = 3)
df_n2_test2_rmse_round <- end_2022_rmse_df$.estimate %>% 
  round(digits = 3)

df_n2_test_fig <- df_n2_test_predictions %>% 
  ggplot(aes(x=week)) +
  geom_line(aes(y=prediction), color = "#FB6A4A",
            linewidth = 2, alpha = .5) +
  geom_point(aes(y=prediction), color = "#FB6A4A", 
             shape = 17, size = 2) +
  geom_ribbon(aes(ymin=pred_lower, ymax=pred_upper), fill = "#FB6A4A", alpha = 0.3) +
  geom_line(aes(y=actual), color = "#003e99",
            linewidth = 2, alpha = .5) +
  geom_point(aes(y=actual), color = "#003e99",
             size = 2) +
  annotate("text", label=df_n2_test1_rsq_round, x=19095, y=1500, size=4) +
  annotate("text", label=df_n2_test2_rsq_round, x=19330, y=1500, size=4) +
  annotate("text", label=df_n2_test1_rmse_round, x=19095, y=1600, size=4) +
  annotate("text", label=df_n2_test2_rmse_round, x=19330, y=1600, size=4) +
  annotate("text", label="Predicted", x=19045, y=1425, color="#FB6A4A", size=4) +
  annotate("text", label="Observed", x=19045, y=1500, color="#003e99", size=4) +
  #ylab("Weekly reported cases") +
  #xlab("Week") +
  geom_vline(xintercept = 19115, linetype = 2) +
  #geom_hline(yintercept = 6800, alpha = 0) +
  ggthemes::theme_clean() +
  theme(axis.title.x = element_blank(),
        axis.title.y = element_blank()) 

df_n2_test_fig

ggsave(here("figures/random_forest/time_series/df_n2_test.png"))

saveRDS(df_n2_test_fig,
        here("figures/random_forest/time_series/df_n2_test.rds"))
```

## Variable importance

```{r}
all_variables <- c("Plant A N1 VL","Plant A N1 DF","Plant A N2 VL",
                  "Plant A N2 DF","Plant B N1 VL","Plant B N1 DF",
                  "Plant B N2 VL","Plant B N2 DF","Plant C N1 VL",
                  "Plant C N1 DF","Plant C N2 VL","Plant C N2 DF")

redon_pal <- c("#5B859EFF", "#1E395FFF", "#75884BFF", "#1E5A46FF", "#DF8D71FF", "#AF4F2FFF", "#D48F90FF", "#732F30FF", "#AB84A5FF", "#59385CFF", "#D8B847FF", "#B38711FF")

# Extract the last 6 colors from the "Reds", "Greens", and "Blues" palettes
reds <- brewer.pal(9, "Reds")[6:9]
greens <- brewer.pal(9, "Greens")[6:9]
blues <- brewer.pal(9, "Blues")[6:9]

# Combine the colors into a single list
color_palette <- c(reds, greens, blues)

vip::vip(all_n6_train_fit, 
         mapping = aes(fill = .data[["Variable"]]),
         num_features = 12) +
  ggthemes::theme_clean() +
  scale_fill_manual(values = color_palette, labels = all_variables) +
  theme(axis.text.y = element_blank(),
        axis.ticks.y = element_blank(),
        #legend.position = "none",
        axis.title.x = element_text(size = 12))

ggsave(here("figures/random_forest/variable_importance/vip_all_predictors.png"),
       width = 6, height = 6.5)
```

```{r}
# Combine the colors into a single list
color_palette2 <- c("#EF3B2C", "#A50F15", "#41AB5D", "#006D2C","#4292C6", "#08519C")

vip::vip(df_n6_train_fit, 
         mapping = aes(fill = .data[["Variable"]]),
         num_features = 6) +
  ggthemes::theme_clean() +
  scale_fill_manual(values = color_palette2) +
  theme(axis.text.y = element_blank(),
        axis.ticks.y = element_blank(),
        #legend.position = "none",
        axis.title.x = element_text(size = 12))

ggsave(here("figures/random_forest/variable_importance/vip_df_predictors.png"),
       width = 6, height = 6.5)
```

```{r}
vip::vip(vl_n6_train_fit, 
         mapping = aes(fill = .data[["Variable"]]),
         num_features = 6) +
  ggthemes::theme_clean() +
  scale_fill_manual(values = color_palette2) +
  theme(axis.text.y = element_blank(),
        axis.ticks.y = element_blank(),
        #legend.position = "none",
        axis.title.x = element_text(size = 12))

ggsave(here("figures/random_forest/variable_importance/vip_vl_predictors.png"),
       width = 6, height = 6.5)
```

```{r}
importance <- vi(all_n6_train_fit) %>% 
  arrange(desc(Importance))

importance
```

```{r}
importance_clean <- importance %>% 
  mutate(plant = str_sub(Variable, 1, 1),
         target = str_sub(Variable, 3, 4),
         data_type = if_else(str_detect(Variable, "POS"), "DF", "VL")) %>% 
  select(!Variable)

importance_clean
```

```{r}
#average importance by plant
plant_vip <- importance_clean %>% 
  group_by(plant) %>% 
  summarise(avg = mean(Importance)) %>% 
  arrange(desc(avg))

plant_vip

A <- plant_vip$avg[1]
B <- plant_vip$avg[2]
C <- plant_vip$avg[3]

A/B
A/C
B/C
```

```{r}
#average importance by plant and gene
plant_gene_vip <- importance_clean %>% 
  group_by(plant,target) %>% 
  summarise(avg = mean(Importance)) %>% 
  arrange(desc(avg))

plant_gene_vip
```

```{r}
#average by plant and data type
importance_clean %>% 
  group_by(plant,data_type) %>% 
  summarise(avg = mean(Importance)) %>% 
  arrange(desc(avg))
```

```{r}
#average by target
gene_vip <- importance_clean %>% 
  group_by(target) %>% 
  summarise(avg = mean(Importance)) %>% 
  arrange(desc(avg))

gene_vip

N1 <- gene_vip$avg[1]
N2 <- gene_vip$avg[2]

N1/N2
```

```{r}
#average by data type
data_vip <- importance_clean %>% 
  group_by(data_type) %>% 
  summarise(avg = mean(Importance)) %>% 
  arrange(desc(avg))

data_vip

DF <- data_vip$avg[1]
VL <- data_vip$avg[2]

DF/VL
```

```{r}
importance_clean %>% 
  group_by(target,data_type) %>% 
  summarise(sum = sum(Importance)) %>% 
  arrange(desc(sum))
```
