==> code/analysis_final/12_RF_LM_compare.Rmd <==
---
title: "RF and LM Comparisons"
author: "Leah Lariscy"
output: html_document
---

# Load packages

```{r}
library(tidyverse)
library(here)
library(RColorBrewer)
library(ggpubr)
library(tsibble)
```

# Load RDS files

```{r}
#all metrics data
lm_metrics <- readRDS(
  here("data/processed_data/linear_reg/metrics_compare.rds"))
rf_metrics <- readRDS(
  here("data/processed_data/rand_forest/metrics_compare.rds"))

#RMSE figures
rf_rmse <- readRDS(here("figures/random_forest/metrics/rmse_compare.rds"))
lm_rmse <- readRDS(here("figures/linear_reg/metrics/rmse_compare.rds"))

#RSQ figures
rf_rsq <- readRDS(here("figures/random_forest/metrics/rsq_compare.rds"))
lm_rsq <- readRDS(here("figures/linear_reg/metrics/rsq_compare.rds"))

#RSQ numbers
rf_rsq_nums <- readRDS(
  here("data/processed_data/rand_forest/rsq_compare.rds"))
lm_rsq_nums <- readRDS(
  here("data/processed_data/linear_reg/rsq_compare.rds"))

#RMSE numbers
rf_rmse_nums <- readRDS(
  here("data/processed_data/rand_forest/rmse_compare.rds"))
lm_rmse_nums <- readRDS(
  here("data/processed_data/linear_reg/rmse_compare.rds"))

#timeseries figures (training data)
rf_all_train <- readRDS(
  here("figures/random_forest/time_series/all_n6_train.rds"))
rf_df_train <- readRDS(
  here("figures/random_forest/time_series/df_n3_train.rds"))
lm_all_train <- readRDS(
  here("figures/linear_reg/time_series/all_n6_train.rds"))
lm_df_train <- readRDS(here(
  "figures/linear_reg/time_series/df_n3_train.rds"))

#timeseries figures (new data)
rf_all_test <- readRDS(
  here("figures/random_forest/time_series/all_n2_test.rds"))
rf_df_test <- readRDS(
  here("figures/random_forest/time_series/df_n2_test.rds"))
lm_all_test <- readRDS(
  here("figures/linear_reg/time_series/all_n2_test.rds"))
lm_df_test <- readRDS(here(
  "figures/linear_reg/time_series/df_n2_test.rds"))
```

# Summary table

```{r}
metric_summary <- rbind(lm_metrics,rf_metrics)

metric_summary

write_csv(metric_summary,
          here("data/processed_data/model_compare/model_metrics.csv"))

metric_summary %>% 
  select(!n,config) %>% 
  group_by(model,feature) %>% 
  summarize_all(mean)
```

# Combine figures

## RMSE

```{r}
ggarrange(rf_rmse,lm_rmse,
          labels = c("a","b"),
          ncol = 2)

ggsave(here("figures/rf_lm_compare/rmse_compare.png"),
            width=14, height=4)
```

```{r}
rf_lm_rmse <- bind_rows(rf_rmse_nums,lm_rmse_nums) %>% 
  arrange(avg_rmse)

rf_lm_rmse
```

```{r}
rf_lm_rmse %>% 
  group_by(feature,model) %>% 
  summarise(mean_rmse = mean(avg_rmse)) %>% 
  arrange(mean_rmse)
```

## RSQ

```{r}
ggarrange(rf_rsq,lm_rsq,
          labels = c("a","b"),
          ncol = 2)

ggsave(here("figures/rf_lm_compare/rsq_compare.png"),
            width=14, height=4)
```

```{r}
rf_lm_rsq <- bind_rows(rf_rsq_nums,lm_rsq_nums) %>% 
  arrange(desc(avg_rsq))

rf_lm_rsq
```

```{r}
rf_lm_rsq %>% 
  group_by(feature,model) %>% 
  summarise(mean_rsq = mean(avg_rsq)) %>% 
  arrange(desc(mean_rsq))
```

## Timeseries - training

```{r}
ggarrange(rf_all_train,rf_df_train,lm_all_train,lm_df_train,
          labels = c("A","B","C","D"),
          ncol = 2,
          nrow = 2)

ggsave(here("figures/rf_lm_compare/training_timeseries.png"),
       width = 12, height = 8)
```

## Timeseries - new data

```{r}
ggarrange(rf_all_test,rf_df_test,lm_all_test,lm_df_test,
          labels = c("a","b","c","d"),
          ncol = 2,
          nrow = 2)

ggsave(here("figures/rf_lm_compare/testing_timeseries.png"),
       width = 12, height = 8)
```
